Finished

+++ hfile/utf8len.h
@@ -1,7 +1,6 @@
 #ifndef UTF8LEN
 #define UTF8LEN
-#include <stddef.h>
 
-size_t utf8len(unsigned char *str);
+signed long long utf8len(unsigned char *str);
 
 #endif
